requirements.txt
Flask_RESTful==0.3.10
flask_sqlalchemy==3.1.1
python-dotenv==1.0.1

flask~=3.1.0
requests~=2.32.3